fix: increase tests
